feat: applied feedback
